added pop-up dialog box
