+++ .eslintrc.js
@@ -22,5 +22,6 @@ module.exports = {
     'react/prop-types': 0,
     'react/jsx-props-no-spreading': 0,
     'linebreak-style': 0,
+    'no-unused-expressions': 0,
   },
 };
