+++ components/explore/SearchResult.module.scss
@@ -1,0 +1,60 @@
+.card {
+	min-height: 60vh;
+	height: auto;
+	width: 25%;
+	margin-left: 2.5%;
+	margin-right: 2.5%;
+	margin-bottom: 4%;
+	border-radius: 7px;
+	overflow: hidden;
+	background-color: rgb(2, 36, 36);
+}
+
+.card img {
+	height: auto;
+	width: 100%;
+	object-fit: cover;
+}
+
+.dish_info {
+	position: relative;
+	padding: 3%;
+}
+
+.dish_name {
+	position: relative;
+	width: 100%;
+	min-height: 5vh;
+	margin-bottom: 1.5vh;
+	height: auto;
+}
+
+.dish_name p {
+	left: 0;
+	width: 60%;
+	font-size: 20px;
+	position: relative;
+}
+.dish_name button {
+	color: rgb(160, 160, 160);
+	top: 0;
+	right: 0;
+	width: 35%;
+	font-size: 15px;
+	height: 4vh;
+	position: absolute;
+	border-style: none;
+	background-color: rgb(65, 65, 65);
+}
+
+.dish_name button:hover {
+	border-style: none;
+	background-color: rgb(86, 86, 86);
+}
+
+.dish_numeric_info p {
+	color: rgb(112, 131, 147);
+	font-size: 15px;
+	width: 100%;
+	height: auto;
+}
