--- components/recipe/SearchResult.module.scss
@@ -1,35 +1,0 @@
-.card {
-	min-height: 60vh;
-	height: auto;
-	width: 25%;
-	margin-left: 2.5%;
-	margin-right: 2.5%;
-	margin-bottom: 4%;
-	border-radius: 7px;
-	overflow: hidden;
-	background-color: rgb(2, 36, 36);
-}
-
-.card img {
-	height: auto;
-	width: 100%;
-	object-fit: cover;
-}
-
-.dish_info {
-	position: relative;
-	padding: 3%;
-}
-
-.dish_name {
-	width: 100%;
-	height: auto;
-	display: flex;
-	flex-direction: row;
-	justify-content: space-between;
-}
-
-.dish_numeric_info {
-	width: 100%;
-	height: auto;
-}
