+++ package.json
@@ -9,6 +9,9 @@
 		"lint": "next lint"
 	},
 	"dependencies": {
+		"@emotion/react": "^11.10.5",
+		"@emotion/styled": "^11.10.5",
+		"@mui/material": "^5.11.4",
 		"@next/font": "13.1.1",
 		"@tanstack/react-query": "^4.20.9",
 		"@tanstack/react-query-devtools": "^4.22.0",
