+++ pages/explore/preloadedData.js
@@ -1,1448 +1,4 @@
 const preloadedData = [
-
-  {
-    recipe: {
-      uri: 'http://www.edamam.com/ontologies/edamam.owl#recipe_e4456795dfe1409f69629df1670a6494',
-      label: 'Fruits Sandwich',
-      image: 'https://edamam-product-images.s3.amazonaws.com/web-img/5d7/5d77596f27841baeb6704eabb4a2fb5c.jpg?X-Amz-Security-Token=IQoJb3JpZ2luX2VjEE0aCXVzLWVhc3QtMSJIMEYCIQCjz1haqH%2BHU8PeUD5XPfPvhrg4B1tIA%2FchIC%2FdqMx%2FuAIhAPgRePErrrJffplWbN5DmRf%2BAVo4D9Hqt1Dj88GSmKD8KtUECKX%2F%2F%2F%2F%2F%2F%2F%2F%2F%2FwEQABoMMTg3MDE3MTUwOTg2Igyr1kArPKYLDQL7Tm4qqQS1JuqF6P2nantseLvzE7MeuUpQfcpjy%2BomgC7n9frUeX32ID6mQFtv6Q0vqUMOYVmLsS4xePgKLGKTLFvOoUUtpmmFiZy187eodT2KgAwjY%2BDU7WWfuYsnCl1PeWwkNTQEkD5O%2Bk3%2FjooadTuzHHLycTfvxJjpoxU5uyA5eYZtrqcPiA1XRG7ObdsEP%2FEuZYPG52mo%2F3oLWlWSpK41%2FrRQQkO4gKMHqriUvmo6YC02bi%2BSZhiAzV1R4C%2FaaW1NuAE%2FULEkSPsFbmyfBaureg6DinmQZlAj7XyP6vfMsUeZJJcs%2BW07eiylvgVi2SxqBWapl74KCeWUMsDZgobtyj6fVXltsjMAQgaZEY3dhmUipBfsXjAruutI0IcVguEIOremf4nSMM6bHGo1i%2BWcTMswT6HTEpDI0bQi%2Bi%2F0MBag8aa%2BE1CJn%2F3lbecKsngM3pvmqOvwAGLtfvzc6zwcBaW6fk%2BIzGXy%2FapYMI3LJjcSZremPwqnUhnjbLtL8DXvN09d1PT8DwXrtFzLkr5%2FervOO8A452AGZniBMqw2Sg0YoZ2Clu8PtxbUAuSy8asYI7Yeu3dsKor%2B1aqkr7b%2BAqALTpNUAuYsE7eZQKR2L62uoUFLu8BNfF5vNsiJV3iLMy%2BCdSAr2zMVnhvUPNK2Wkb0dfxZ7t89bznZVuH5ZBQ2Oh1TB9Ns8BjBD%2BMUaplrhya8IQXNk0npB4D3vHznNvjYVY4%2F%2FSiGZXYkMO7Q%2Bp0GOqgBhUfjEhUQQm93O5R9FZCA7T8ezNbjnvoWDlhSAl6MlnpXCgWuy3hIjNaGk60E4TPv0r7tLkyMiy28ZEtyMPt7iUka%2B2zJIr5pMDvbpFXEYYSpkOO1rl%2FoWE5uTgdEEN%2BWYm%2BtWMizTJIuFJ8e%2BzE%2BBkDxPQG7a3S%2B1pwp3OkKZ5k%2FnuvdyvQ2zlATQFxyFEhImfF3uo2y9RgpZfT8KooBqeL0tg%2FfUt3%2F&X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Date=20230111T131700Z&X-Amz-SignedHeaders=host&X-Amz-Expires=3600&X-Amz-Credential=ASIASXCYXIIFFLFRDZEE%2F20230111%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Signature=0b35c5a51686546c8247c938a7b2098db7aaea0f7f516512ed9d354ec16349e9',
-      source: 'Food52',
-      url: 'https://food52.com/recipes/61017-fruits-sandwich',
-      shareAs: 'http://www.edamam.com/recipe/fruits-sandwich-e4456795dfe1409f69629df1670a6494/sandwich',
-      yield: 2,
-      dietLabels: [
-        'High-Fiber',
-      ],
-      healthLabels: [],
-      cautions: [
-        'Sulfites',
-      ],
-      ingredientLines: [
-        '100 milliliters heavy cream',
-        '1 tablespoon sugar',
-        '4 pieces white sandwich bread',
-        '4 pieces strawberry',
-        '1 piece banana',
-        '2 pieces sliced kiwi',
-        '8 pieces tangerine',
-      ],
-      ingredients: [
-        {
-          text: '100 milliliters heavy cream',
-          quantity: 100,
-          measure: 'milliliter',
-          food: 'heavy cream',
-          weight: 100.76578765149213,
-          foodCategory: 'Dairy',
-          foodId: 'food_bgtkr21b5v16mca246x9ebnaswyo',
-          image: 'https://www.edamam.com/food-img/484/4848d71f6a14dd5076083f5e17925420.jpg',
-        },
-        {
-          text: '1 tablespoon sugar',
-          quantity: 1,
-          measure: 'tablespoon',
-          food: 'sugar',
-          weight: 12.4999999997887,
-          foodCategory: 'sugars',
-          foodId: 'food_axi2ijobrk819yb0adceobnhm1c2',
-          image: 'https://www.edamam.com/food-img/ecb/ecb3f5aaed96d0188c21b8369be07765.jpg',
-        },
-        {
-          text: '4 pieces white sandwich bread',
-          quantity: 4,
-          measure: 'piece',
-          food: 'sandwich bread',
-          weight: 116,
-          foodCategory: 'bread, rolls and tortillas',
-          foodId: 'food_a3049hmbqj5wstaeeb3udaz6uaqv',
-          image: 'https://www.edamam.com/food-img/886/886960f6ce6ccec5b9163bacf2996853.jpg',
-        },
-        {
-          text: '4 pieces strawberry',
-          quantity: 4,
-          measure: '<unit>',
-          food: 'strawberry',
-          weight: 53.733333333333334,
-          foodCategory: 'fruit',
-          foodId: 'food_b4s2ibkbrrucmbabbaxhfau8ay42',
-          image: 'https://www.edamam.com/food-img/00c/00c8851e401bf7975be7f73499b4b573.jpg',
-        },
-        {
-          text: '1 piece banana',
-          quantity: 1,
-          measure: '<unit>',
-          food: 'banana',
-          weight: 115.7,
-          foodCategory: 'fruit',
-          foodId: 'food_bjsfxtcaidvmhaa3afrbna43q3hu',
-          image: 'https://www.edamam.com/food-img/9f6/9f6181163a25c96022ee3fc66d9ebb11.jpg',
-        },
-        {
-          text: '2 pieces sliced kiwi',
-          quantity: 2,
-          measure: '<unit>',
-          food: 'kiwi',
-          weight: 138,
-          foodCategory: 'fruit',
-          foodId: 'food_bqj114aab4xnl6b6j3dt0bi4sznd',
-          image: 'https://www.edamam.com/food-img/dba/dbabbf58b6a43c9b65014f4c789d4409.jpg',
-        },
-        {
-          text: '8 pieces tangerine',
-          quantity: 8,
-          measure: 'piece',
-          food: 'tangerine',
-          weight: 226.4,
-          foodCategory: 'fruit',
-          foodId: 'food_a24i0asarijf6nbnhkbbrbgy91bw',
-          image: 'https://www.edamam.com/food-img/f82/f82193abe21e23c1562bb5340f74e336.jpg',
-        },
-      ],
-      calories: 1030.0766340634966,
-      totalWeight: 763.0991209846142,
-      totalTime: 15,
-      cuisineType: [
-        'british',
-      ],
-      mealType: [
-        'lunch/dinner',
-      ],
-      dishType: [
-        'sandwiches',
-      ],
-      totalNutrients: {
-        ENERC_KCAL: {
-          label: 'Energy',
-          quantity: 1030.0766340634966,
-          unit: 'kcal',
-        },
-        FAT: {
-          label: 'Fat',
-          quantity: 43.004191431052085,
-          unit: 'g',
-        },
-        FASAT: {
-          label: 'Saturated',
-          quantity: 24.375656211891666,
-          unit: 'g',
-        },
-        FATRN: {
-          label: 'Trans',
-          quantity: 0.030159999999999996,
-          unit: 'g',
-        },
-        FAMS: {
-          label: 'Monounsaturated',
-          quantity: 11.732781401771783,
-          unit: 'g',
-        },
-        FAPU: {
-          label: 'Polyunsaturated',
-          quantity: 3.8343295889981683,
-          unit: 'g',
-        },
-        CHOCDF: {
-          label: 'Carbs',
-          quantity: 152.76282547526537,
-          unit: 'g',
-        },
-        'CHOCDF.net': {
-          label: 'Carbohydrates (net)',
-          quantity: 135.82475880859872,
-          unit: 'g',
-        },
-        FIBTG: {
-          label: 'Fiber',
-          quantity: 16.938066666666668,
-          unit: 'g',
-        },
-        SUGAR: {
-          label: 'Sugars',
-          quantity: 75.27895547526575,
-          unit: 'g',
-        },
-        'SUGAR.added': {
-          label: 'Sugars, added',
-          quantity: 12.474999999789123,
-          unit: 'g',
-        },
-        PROCNT: {
-          label: 'Protein',
-          quantity: 19.52908198018892,
-          unit: 'g',
-        },
-        CHOLE: {
-          label: 'Cholesterol',
-          quantity: 138.04912908254423,
-          unit: 'mg',
-        },
-        NA: {
-          label: 'Sodium',
-          quantity: 638.0583326408982,
-          unit: 'mg',
-        },
-        CA: {
-          label: 'Calcium',
-          quantity: 364.9730953068011,
-          unit: 'mg',
-        },
-        MG: {
-          label: 'Magnesium',
-          quantity: 148.10593846893778,
-          unit: 'mg',
-        },
-        K: {
-          label: 'Potassium',
-          quantity: 1583.9463407386152,
-          unit: 'mg',
-        },
-        FE: {
-          label: 'Iron',
-          quantity: 5.373406402962009,
-          unit: 'mg',
-        },
-        ZN: {
-          label: 'Zinc',
-          quantity: 2.2138679782650774,
-          unit: 'mg',
-        },
-        P: {
-          label: 'Phosphorus',
-          quantity: 365.86478834392517,
-          unit: 'mg',
-        },
-        VITA_RAE: {
-          label: 'Vitamin A',
-          quantity: 500.65172058096596,
-          unit: 'µg',
-        },
-        VITC: {
-          label: 'Vitamin C',
-          quantity: 230.87249472590898,
-          unit: 'mg',
-        },
-        THIA: {
-          label: 'Thiamin (B1)',
-          quantity: 0.7209034732833282,
-          unit: 'mg',
-        },
-        RIBF: {
-          label: 'Riboflavin (B2)',
-          quantity: 0.6189836997499345,
-          unit: 'mg',
-        },
-        NIA: {
-          label: 'Niacin (B3)',
-          quantity: 8.857158323850749,
-          unit: 'mg',
-        },
-        VITB6A: {
-          label: 'Vitamin B6',
-          quantity: 0.8683647714560547,
-          unit: 'mg',
-        },
-        FOLDFE: {
-          label: 'Folate equivalent (total)',
-          quantity: 225.63063150605967,
-          unit: 'µg',
-        },
-        FOLFD: {
-          label: 'Folate (food)',
-          quantity: 186.19063150605967,
-          unit: 'µg',
-        },
-        FOLAC: {
-          label: 'Folic acid',
-          quantity: 23.2,
-          unit: 'µg',
-        },
-        VITB12: {
-          label: 'Vitamin B12',
-          quantity: 0.18137841777268585,
-          unit: 'µg',
-        },
-        VITD: {
-          label: 'Vitamin D',
-          quantity: 0.7053605135604449,
-          unit: 'µg',
-        },
-        TOCPHA: {
-          label: 'Vitamin E',
-          quantity: 4.027644015772483,
-          unit: 'mg',
-        },
-        VITK1: {
-          label: 'Vitamin K',
-          quantity: 66.28313853818108,
-          unit: 'µg',
-        },
-        'Sugar.alcohol': {
-          label: 'Sugar alcohol',
-          quantity: 0,
-          unit: 'g',
-        },
-        WATER: {
-          label: 'Water',
-          quantity: 541.9892527203427,
-          unit: 'g',
-        },
-      },
-      totalDaily: {
-        ENERC_KCAL: {
-          label: 'Energy',
-          quantity: 51.50383170317483,
-          unit: '%',
-        },
-        FAT: {
-          label: 'Fat',
-          quantity: 66.1602945093109,
-          unit: '%',
-        },
-        FASAT: {
-          label: 'Saturated',
-          quantity: 121.87828105945832,
-          unit: '%',
-        },
-        CHOCDF: {
-          label: 'Carbs',
-          quantity: 50.92094182508846,
-          unit: '%',
-        },
-        FIBTG: {
-          label: 'Fiber',
-          quantity: 67.75226666666667,
-          unit: '%',
-        },
-        PROCNT: {
-          label: 'Protein',
-          quantity: 39.05816396037784,
-          unit: '%',
-        },
-        CHOLE: {
-          label: 'Cholesterol',
-          quantity: 46.01637636084808,
-          unit: '%',
-        },
-        NA: {
-          label: 'Sodium',
-          quantity: 26.585763860037424,
-          unit: '%',
-        },
-        CA: {
-          label: 'Calcium',
-          quantity: 36.49730953068011,
-          unit: '%',
-        },
-        MG: {
-          label: 'Magnesium',
-          quantity: 35.26331868308043,
-          unit: '%',
-        },
-        K: {
-          label: 'Potassium',
-          quantity: 33.70098597316203,
-          unit: '%',
-        },
-        FE: {
-          label: 'Iron',
-          quantity: 29.85225779423338,
-          unit: '%',
-        },
-        ZN: {
-          label: 'Zinc',
-          quantity: 20.12607252968252,
-          unit: '%',
-        },
-        P: {
-          label: 'Phosphorus',
-          quantity: 52.26639833484646,
-          unit: '%',
-        },
-        VITA_RAE: {
-          label: 'Vitamin A',
-          quantity: 55.62796895344066,
-          unit: '%',
-        },
-        VITC: {
-          label: 'Vitamin C',
-          quantity: 256.52499413989887,
-          unit: '%',
-        },
-        THIA: {
-          label: 'Thiamin (B1)',
-          quantity: 60.075289440277345,
-          unit: '%',
-        },
-        RIBF: {
-          label: 'Riboflavin (B2)',
-          quantity: 47.61413074999496,
-          unit: '%',
-        },
-        NIA: {
-          label: 'Niacin (B3)',
-          quantity: 55.35723952406718,
-          unit: '%',
-        },
-        VITB6A: {
-          label: 'Vitamin B6',
-          quantity: 66.7972901120042,
-          unit: '%',
-        },
-        FOLDFE: {
-          label: 'Folate equivalent (total)',
-          quantity: 56.40765787651491,
-          unit: '%',
-        },
-        VITB12: {
-          label: 'Vitamin B12',
-          quantity: 7.55743407386191,
-          unit: '%',
-        },
-        VITD: {
-          label: 'Vitamin D',
-          quantity: 4.7024034237363,
-          unit: '%',
-        },
-        TOCPHA: {
-          label: 'Vitamin E',
-          quantity: 26.85096010514989,
-          unit: '%',
-        },
-        VITK1: {
-          label: 'Vitamin K',
-          quantity: 55.23594878181756,
-          unit: '%',
-        },
-      },
-      digest: [
-        {
-          label: 'Fat',
-          tag: 'FAT',
-          schemaOrgTag: 'fatContent',
-          total: 43.004191431052085,
-          hasRDI: true,
-          daily: 66.1602945093109,
-          unit: 'g',
-          sub: [
-            {
-              label: 'Saturated',
-              tag: 'FASAT',
-              schemaOrgTag: 'saturatedFatContent',
-              total: 24.375656211891666,
-              hasRDI: true,
-              daily: 121.87828105945832,
-              unit: 'g',
-            },
-            {
-              label: 'Trans',
-              tag: 'FATRN',
-              schemaOrgTag: 'transFatContent',
-              total: 0.030159999999999996,
-              hasRDI: false,
-              daily: 0,
-              unit: 'g',
-            },
-            {
-              label: 'Monounsaturated',
-              tag: 'FAMS',
-              schemaOrgTag: null,
-              total: 11.732781401771783,
-              hasRDI: false,
-              daily: 0,
-              unit: 'g',
-            },
-            {
-              label: 'Polyunsaturated',
-              tag: 'FAPU',
-              schemaOrgTag: null,
-              total: 3.8343295889981683,
-              hasRDI: false,
-              daily: 0,
-              unit: 'g',
-            },
-          ],
-        },
-        {
-          label: 'Carbs',
-          tag: 'CHOCDF',
-          schemaOrgTag: 'carbohydrateContent',
-          total: 152.76282547526537,
-          hasRDI: true,
-          daily: 50.92094182508846,
-          unit: 'g',
-          sub: [
-            {
-              label: 'Carbs (net)',
-              tag: 'CHOCDF.net',
-              schemaOrgTag: null,
-              total: 135.82475880859872,
-              hasRDI: false,
-              daily: 0,
-              unit: 'g',
-            },
-            {
-              label: 'Fiber',
-              tag: 'FIBTG',
-              schemaOrgTag: 'fiberContent',
-              total: 16.938066666666668,
-              hasRDI: true,
-              daily: 67.75226666666667,
-              unit: 'g',
-            },
-            {
-              label: 'Sugars',
-              tag: 'SUGAR',
-              schemaOrgTag: 'sugarContent',
-              total: 75.27895547526575,
-              hasRDI: false,
-              daily: 0,
-              unit: 'g',
-            },
-            {
-              label: 'Sugars, added',
-              tag: 'SUGAR.added',
-              schemaOrgTag: null,
-              total: 12.474999999789123,
-              hasRDI: false,
-              daily: 0,
-              unit: 'g',
-            },
-          ],
-        },
-        {
-          label: 'Protein',
-          tag: 'PROCNT',
-          schemaOrgTag: 'proteinContent',
-          total: 19.52908198018892,
-          hasRDI: true,
-          daily: 39.05816396037784,
-          unit: 'g',
-        },
-        {
-          label: 'Cholesterol',
-          tag: 'CHOLE',
-          schemaOrgTag: 'cholesterolContent',
-          total: 138.04912908254423,
-          hasRDI: true,
-          daily: 46.01637636084808,
-          unit: 'mg',
-        },
-        {
-          label: 'Sodium',
-          tag: 'NA',
-          schemaOrgTag: 'sodiumContent',
-          total: 638.0583326408982,
-          hasRDI: true,
-          daily: 26.585763860037424,
-          unit: 'mg',
-        },
-        {
-          label: 'Calcium',
-          tag: 'CA',
-          schemaOrgTag: null,
-          total: 364.9730953068011,
-          hasRDI: true,
-          daily: 36.49730953068011,
-          unit: 'mg',
-        },
-        {
-          label: 'Magnesium',
-          tag: 'MG',
-          schemaOrgTag: null,
-          total: 148.10593846893778,
-          hasRDI: true,
-          daily: 35.26331868308043,
-          unit: 'mg',
-        },
-        {
-          label: 'Potassium',
-          tag: 'K',
-          schemaOrgTag: null,
-          total: 1583.9463407386152,
-          hasRDI: true,
-          daily: 33.70098597316203,
-          unit: 'mg',
-        },
-        {
-          label: 'Iron',
-          tag: 'FE',
-          schemaOrgTag: null,
-          total: 5.373406402962009,
-          hasRDI: true,
-          daily: 29.85225779423338,
-          unit: 'mg',
-        },
-        {
-          label: 'Zinc',
-          tag: 'ZN',
-          schemaOrgTag: null,
-          total: 2.2138679782650774,
-          hasRDI: true,
-          daily: 20.12607252968252,
-          unit: 'mg',
-        },
-        {
-          label: 'Phosphorus',
-          tag: 'P',
-          schemaOrgTag: null,
-          total: 365.86478834392517,
-          hasRDI: true,
-          daily: 52.26639833484646,
-          unit: 'mg',
-        },
-        {
-          label: 'Vitamin A',
-          tag: 'VITA_RAE',
-          schemaOrgTag: null,
-          total: 500.65172058096596,
-          hasRDI: true,
-          daily: 55.62796895344066,
-          unit: 'µg',
-        },
-        {
-          label: 'Vitamin C',
-          tag: 'VITC',
-          schemaOrgTag: null,
-          total: 230.87249472590898,
-          hasRDI: true,
-          daily: 256.52499413989887,
-          unit: 'mg',
-        },
-        {
-          label: 'Thiamin (B1)',
-          tag: 'THIA',
-          schemaOrgTag: null,
-          total: 0.7209034732833282,
-          hasRDI: true,
-          daily: 60.075289440277345,
-          unit: 'mg',
-        },
-        {
-          label: 'Riboflavin (B2)',
-          tag: 'RIBF',
-          schemaOrgTag: null,
-          total: 0.6189836997499345,
-          hasRDI: true,
-          daily: 47.61413074999496,
-          unit: 'mg',
-        },
-        {
-          label: 'Niacin (B3)',
-          tag: 'NIA',
-          schemaOrgTag: null,
-          total: 8.857158323850749,
-          hasRDI: true,
-          daily: 55.35723952406718,
-          unit: 'mg',
-        },
-        {
-          label: 'Vitamin B6',
-          tag: 'VITB6A',
-          schemaOrgTag: null,
-          total: 0.8683647714560547,
-          hasRDI: true,
-          daily: 66.7972901120042,
-          unit: 'mg',
-        },
-        {
-          label: 'Folate equivalent (total)',
-          tag: 'FOLDFE',
-          schemaOrgTag: null,
-          total: 225.63063150605967,
-          hasRDI: true,
-          daily: 56.40765787651491,
-          unit: 'µg',
-        },
-        {
-          label: 'Folate (food)',
-          tag: 'FOLFD',
-          schemaOrgTag: null,
-          total: 186.19063150605967,
-          hasRDI: false,
-          daily: 0,
-          unit: 'µg',
-        },
-        {
-          label: 'Folic acid',
-          tag: 'FOLAC',
-          schemaOrgTag: null,
-          total: 23.2,
-          hasRDI: false,
-          daily: 0,
-          unit: 'µg',
-        },
-        {
-          label: 'Vitamin B12',
-          tag: 'VITB12',
-          schemaOrgTag: null,
-          total: 0.18137841777268585,
-          hasRDI: true,
-          daily: 7.55743407386191,
-          unit: 'µg',
-        },
-        {
-          label: 'Vitamin D',
-          tag: 'VITD',
-          schemaOrgTag: null,
-          total: 0.7053605135604449,
-          hasRDI: true,
-          daily: 4.7024034237363,
-          unit: 'µg',
-        },
-        {
-          label: 'Vitamin E',
-          tag: 'TOCPHA',
-          schemaOrgTag: null,
-          total: 4.027644015772483,
-          hasRDI: true,
-          daily: 26.85096010514989,
-          unit: 'mg',
-        },
-        {
-          label: 'Vitamin K',
-          tag: 'VITK1',
-          schemaOrgTag: null,
-          total: 66.28313853818108,
-          hasRDI: true,
-          daily: 55.23594878181756,
-          unit: 'µg',
-        },
-        {
-          label: 'Sugar alcohols',
-          tag: 'Sugar.alcohol',
-          schemaOrgTag: null,
-          total: 0,
-          hasRDI: false,
-          daily: 0,
-          unit: 'g',
-        },
-        {
-          label: 'Water',
-          tag: 'WATER',
-          schemaOrgTag: null,
-          total: 541.9892527203427,
-          hasRDI: false,
-          daily: 0,
-          unit: 'g',
-        },
-      ],
-    },
-  },
-  {
-    recipe: {
-      uri: 'http://www.edamam.com/ontologies/edamam.owl#recipe_2256482f9dc843a41d747ca9d76690e5',
-      label: 'Turkey Club Sandwich recipes',
-      image: 'https://edamam-product-images.s3.amazonaws.com/web-img/ab5/ab5ddc22f4c01397fca3c24e2dab0617?X-Amz-Security-Token=IQoJb3JpZ2luX2VjEE0aCXVzLWVhc3QtMSJIMEYCIQCjz1haqH%2BHU8PeUD5XPfPvhrg4B1tIA%2FchIC%2FdqMx%2FuAIhAPgRePErrrJffplWbN5DmRf%2BAVo4D9Hqt1Dj88GSmKD8KtUECKX%2F%2F%2F%2F%2F%2F%2F%2F%2F%2FwEQABoMMTg3MDE3MTUwOTg2Igyr1kArPKYLDQL7Tm4qqQS1JuqF6P2nantseLvzE7MeuUpQfcpjy%2BomgC7n9frUeX32ID6mQFtv6Q0vqUMOYVmLsS4xePgKLGKTLFvOoUUtpmmFiZy187eodT2KgAwjY%2BDU7WWfuYsnCl1PeWwkNTQEkD5O%2Bk3%2FjooadTuzHHLycTfvxJjpoxU5uyA5eYZtrqcPiA1XRG7ObdsEP%2FEuZYPG52mo%2F3oLWlWSpK41%2FrRQQkO4gKMHqriUvmo6YC02bi%2BSZhiAzV1R4C%2FaaW1NuAE%2FULEkSPsFbmyfBaureg6DinmQZlAj7XyP6vfMsUeZJJcs%2BW07eiylvgVi2SxqBWapl74KCeWUMsDZgobtyj6fVXltsjMAQgaZEY3dhmUipBfsXjAruutI0IcVguEIOremf4nSMM6bHGo1i%2BWcTMswT6HTEpDI0bQi%2Bi%2F0MBag8aa%2BE1CJn%2F3lbecKsngM3pvmqOvwAGLtfvzc6zwcBaW6fk%2BIzGXy%2FapYMI3LJjcSZremPwqnUhnjbLtL8DXvN09d1PT8DwXrtFzLkr5%2FervOO8A452AGZniBMqw2Sg0YoZ2Clu8PtxbUAuSy8asYI7Yeu3dsKor%2B1aqkr7b%2BAqALTpNUAuYsE7eZQKR2L62uoUFLu8BNfF5vNsiJV3iLMy%2BCdSAr2zMVnhvUPNK2Wkb0dfxZ7t89bznZVuH5ZBQ2Oh1TB9Ns8BjBD%2BMUaplrhya8IQXNk0npB4D3vHznNvjYVY4%2F%2FSiGZXYkMO7Q%2Bp0GOqgBhUfjEhUQQm93O5R9FZCA7T8ezNbjnvoWDlhSAl6MlnpXCgWuy3hIjNaGk60E4TPv0r7tLkyMiy28ZEtyMPt7iUka%2B2zJIr5pMDvbpFXEYYSpkOO1rl%2FoWE5uTgdEEN%2BWYm%2BtWMizTJIuFJ8e%2BzE%2BBkDxPQG7a3S%2B1pwp3OkKZ5k%2FnuvdyvQ2zlATQFxyFEhImfF3uo2y9RgpZfT8KooBqeL0tg%2FfUt3%2F&X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Date=20230111T131700Z&X-Amz-SignedHeaders=host&X-Amz-Expires=3600&X-Amz-Credential=ASIASXCYXIIFFLFRDZEE%2F20230111%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Signature=36e24c9d9e86383017fae3d366ea0bdcf79f46668d2f8a0cd7e958c41c085b69',
-      source: 'Food Republic',
-      url: 'http://www.foodrepublic.com/recipes/turkey-club-sandwich-recipe/',
-      shareAs: 'http://www.edamam.com/recipe/turkey-club-sandwich-recipes-2256482f9dc843a41d747ca9d76690e5/sandwich',
-      yield: 4,
-      dietLabels: [
-        'Low-Carb',
-      ],
-      healthLabels: [],
-      cautions: [
-        'Sulfites',
-      ],
-      ingredientLines: [
-        '3 slices white sandwich bread, lightly toasted',
-        '2 tablespoons mayonnaise',
-        '1/2 pound turkey breast, thinly sliced',
-        '4 slices thick-cut bacon, fried until crisp',
-        '1 small tomato, sliced into rounds',
-        '4 leaves lettuce, rinsed and dried',
-      ],
-      ingredients: [
-        {
-          text: '3 slices white sandwich bread, lightly toasted',
-          quantity: 3,
-          measure: 'slice',
-          food: 'sandwich bread',
-          weight: 87,
-          foodCategory: 'bread, rolls and tortillas',
-          foodId: 'food_a3049hmbqj5wstaeeb3udaz6uaqv',
-          image: 'https://www.edamam.com/food-img/886/886960f6ce6ccec5b9163bacf2996853.jpg',
-        },
-        {
-          text: '2 tablespoons mayonnaise',
-          quantity: 2,
-          measure: 'tablespoon',
-          food: 'mayonnaise',
-          weight: 28.8749999995118,
-          foodCategory: 'condiments and sauces',
-          foodId: 'food_bu8t61zaplle7dbrzk81dbygq0qj',
-          image: 'https://www.edamam.com/food-img/577/577308a0422357885c94cc9b5f1f1862.jpg',
-        },
-        {
-          text: '1/2 pound turkey breast, thinly sliced',
-          quantity: 0.5,
-          measure: 'pound',
-          food: 'turkey breast',
-          weight: 226.796185,
-          foodCategory: 'Poultry',
-          foodId: 'food_aijy334arptsr3b4zwpdrajqapl8',
-          image: 'https://www.edamam.com/food-img/3d8/3d80f3f4fcb0efcf54d094c857d74926.jpg',
-        },
-        {
-          text: '4 slices thick-cut bacon, fried until crisp',
-          quantity: 4,
-          measure: 'slice',
-          food: 'bacon',
-          weight: 116,
-          foodCategory: 'cured meats',
-          foodId: 'food_av0awb0bgs18qub01sd9kbx44b61',
-          image: 'https://www.edamam.com/food-img/d42/d426884a125fa39a70d5a5d7217864ec.jpg',
-        },
-        {
-          text: '1 small tomato, sliced into rounds',
-          quantity: 1,
-          measure: '<unit>',
-          food: 'tomato',
-          weight: 92.25,
-          foodCategory: 'vegetables',
-          foodId: 'food_a6k79rrahp8fe2b26zussa3wtkqh',
-          image: 'https://www.edamam.com/food-img/23e/23e727a14f1035bdc2733bb0477efbd2.jpg',
-        },
-        {
-          text: '4 leaves lettuce, rinsed and dried',
-          quantity: 4,
-          measure: 'leaf',
-          food: 'lettuce',
-          weight: 20,
-          foodCategory: 'vegetables',
-          foodId: 'food_bf5fxtkbc9alwoajuvsi7amonr5w',
-          image: 'https://www.edamam.com/food-img/719/71996625d0cb47e197093ecd52c97dc2.jpg',
-        },
-      ],
-      calories: 1291.1000104466216,
-      totalWeight: 570.9211849995118,
-      totalTime: 15,
-      cuisineType: [
-        'american',
-      ],
-      mealType: [
-        'lunch/dinner',
-      ],
-      dishType: [
-        'sandwiches',
-      ],
-      totalNutrients: {
-        ENERC_KCAL: {
-          label: 'Energy',
-          quantity: 1291.1000104466216,
-          unit: 'kcal',
-        },
-        FAT: {
-          label: 'Fat',
-          quantity: 87.24254218662408,
-          unit: 'g',
-        },
-        FASAT: {
-          label: 'Saturated',
-          quantity: 23.783412133443857,
-          unit: 'g',
-        },
-        FATRN: {
-          label: 'Trans',
-          quantity: 0.1769,
-          unit: 'g',
-        },
-        FAMS: {
-          label: 'Monounsaturated',
-          quantity: 32.39318102090577,
-          unit: 'g',
-        },
-        FAPU: {
-          label: 'Polyunsaturated',
-          quantity: 25.995804170774456,
-          unit: 'g',
-        },
-        CHOCDF: {
-          label: 'Carbs',
-          quantity: 47.87092499999999,
-          unit: 'g',
-        },
-        'CHOCDF.net': {
-          label: 'Carbohydrates (net)',
-          quantity: 43.06392499999999,
-          unit: 'g',
-        },
-        FIBTG: {
-          label: 'Fiber',
-          quantity: 4.8069999999999995,
-          unit: 'g',
-        },
-        SUGAR: {
-          label: 'Sugars',
-          quantity: 8.915875000000002,
-          unit: 'g',
-        },
-        'SUGAR.added': {
-          label: 'Sugars, added',
-          quantity: 0,
-          unit: 'g',
-        },
-        PROCNT: {
-          label: 'Protein',
-          quantity: 74.69308489650001,
-          unit: 'g',
-        },
-        CHOLE: {
-          label: 'Cholesterol',
-          quantity: 235.2387702498096,
-          unit: 'mg',
-        },
-        NA: {
-          label: 'Sodium',
-          quantity: 1515.910999147183,
-          unit: 'mg',
-        },
-        CA: {
-          label: 'Calcium',
-          quantity: 167.21850405,
-          unit: 'mg',
-        },
-        MG: {
-          label: 'Magnesium',
-          quantity: 120.2485844,
-          unit: 'mg',
-        },
-        K: {
-          label: 'Potassium',
-          quantity: 1273.5920087499999,
-          unit: 'mg',
-        },
-        FE: {
-          label: 'Iron',
-          quantity: 6.730529220000001,
-          unit: 'mg',
-        },
-        ZN: {
-          label: 'Zinc',
-          quantity: 6.161625104500001,
-          unit: 'mg',
-        },
-        P: {
-          label: 'Phosphorus',
-          quantity: 747.2509041000001,
-          unit: 'mg',
-        },
-        VITA_RAE: {
-          label: 'Vitamin A',
-          quantity: 89.2409237,
-          unit: 'µg',
-        },
-        VITC: {
-          label: 'Vitamin C',
-          quantity: 13.552249999999999,
-          unit: 'mg',
-        },
-        THIA: {
-          label: 'Thiamin (B1)',
-          quantity: 0.8582842873,
-          unit: 'mg',
-        },
-        RIBF: {
-          label: 'Riboflavin (B2)',
-          quantity: 0.60481311275,
-          unit: 'mg',
-        },
-        NIA: {
-          label: 'Niacin (B3)',
-          quantity: 21.967686620000006,
-          unit: 'mg',
-        },
-        VITB6A: {
-          label: 'Vitamin B6',
-          quantity: 1.5839516880000002,
-          unit: 'mg',
-        },
-        FOLDFE: {
-          label: 'Folate equivalent (total)',
-          quantity: 130.44323295,
-          unit: 'µg',
-        },
-        FOLFD: {
-          label: 'Folate (food)',
-          quantity: 100.86323295,
-          unit: 'µg',
-        },
-        FOLAC: {
-          label: 'Folic acid',
-          quantity: 17.4,
-          unit: 'µg',
-        },
-        VITB12: {
-          label: 'Vitamin B12',
-          quantity: 1.532543977,
-          unit: 'µg',
-        },
-        VITD: {
-          label: 'Vitamin D',
-          quantity: 0.46399999999999997,
-          unit: 'µg',
-        },
-        TOCPHA: {
-          label: 'Vitamin E',
-          quantity: 1.19825,
-          unit: 'mg',
-        },
-        VITK1: {
-          label: 'Vitamin K',
-          quantity: 32.01075,
-          unit: 'µg',
-        },
-        'Sugar.alcohol': {
-          label: 'Sugar alcohol',
-          quantity: 0,
-          unit: 'g',
-        },
-        WATER: {
-          label: 'Water',
-          quantity: 347.1338275925,
-          unit: 'g',
-        },
-      },
-      totalDaily: {
-        ENERC_KCAL: {
-          label: 'Energy',
-          quantity: 64.55500052233108,
-          unit: '%',
-        },
-        FAT: {
-          label: 'Fat',
-          quantity: 134.21929567172936,
-          unit: '%',
-        },
-        FASAT: {
-          label: 'Saturated',
-          quantity: 118.91706066721929,
-          unit: '%',
-        },
-        CHOCDF: {
-          label: 'Carbs',
-          quantity: 15.956974999999996,
-          unit: '%',
-        },
-        FIBTG: {
-          label: 'Fiber',
-          quantity: 19.227999999999998,
-          unit: '%',
-        },
-        PROCNT: {
-          label: 'Protein',
-          quantity: 149.38616979300002,
-          unit: '%',
-        },
-        CHOLE: {
-          label: 'Cholesterol',
-          quantity: 78.4129234166032,
-          unit: '%',
-        },
-        NA: {
-          label: 'Sodium',
-          quantity: 63.16295829779929,
-          unit: '%',
-        },
-        CA: {
-          label: 'Calcium',
-          quantity: 16.721850405,
-          unit: '%',
-        },
-        MG: {
-          label: 'Magnesium',
-          quantity: 28.630615333333335,
-          unit: '%',
-        },
-        K: {
-          label: 'Potassium',
-          quantity: 27.09770231382978,
-          unit: '%',
-        },
-        FE: {
-          label: 'Iron',
-          quantity: 37.391829,
-          unit: '%',
-        },
-        ZN: {
-          label: 'Zinc',
-          quantity: 56.01477367727273,
-          unit: '%',
-        },
-        P: {
-          label: 'Phosphorus',
-          quantity: 106.75012915714287,
-          unit: '%',
-        },
-        VITA_RAE: {
-          label: 'Vitamin A',
-          quantity: 9.915658188888889,
-          unit: '%',
-        },
-        VITC: {
-          label: 'Vitamin C',
-          quantity: 15.058055555555555,
-          unit: '%',
-        },
-        THIA: {
-          label: 'Thiamin (B1)',
-          quantity: 71.52369060833334,
-          unit: '%',
-        },
-        RIBF: {
-          label: 'Riboflavin (B2)',
-          quantity: 46.524085596153846,
-          unit: '%',
-        },
-        NIA: {
-          label: 'Niacin (B3)',
-          quantity: 137.29804137500003,
-          unit: '%',
-        },
-        VITB6A: {
-          label: 'Vitamin B6',
-          quantity: 121.84243753846155,
-          unit: '%',
-        },
-        FOLDFE: {
-          label: 'Folate equivalent (total)',
-          quantity: 32.6108082375,
-          unit: '%',
-        },
-        VITB12: {
-          label: 'Vitamin B12',
-          quantity: 63.85599904166667,
-          unit: '%',
-        },
-        VITD: {
-          label: 'Vitamin D',
-          quantity: 3.0933333333333333,
-          unit: '%',
-        },
-        TOCPHA: {
-          label: 'Vitamin E',
-          quantity: 7.988333333333333,
-          unit: '%',
-        },
-        VITK1: {
-          label: 'Vitamin K',
-          quantity: 26.675625000000004,
-          unit: '%',
-        },
-      },
-      digest: [
-        {
-          label: 'Fat',
-          tag: 'FAT',
-          schemaOrgTag: 'fatContent',
-          total: 87.24254218662408,
-          hasRDI: true,
-          daily: 134.21929567172936,
-          unit: 'g',
-          sub: [
-            {
-              label: 'Saturated',
-              tag: 'FASAT',
-              schemaOrgTag: 'saturatedFatContent',
-              total: 23.783412133443857,
-              hasRDI: true,
-              daily: 118.91706066721929,
-              unit: 'g',
-            },
-            {
-              label: 'Trans',
-              tag: 'FATRN',
-              schemaOrgTag: 'transFatContent',
-              total: 0.1769,
-              hasRDI: false,
-              daily: 0,
-              unit: 'g',
-            },
-            {
-              label: 'Monounsaturated',
-              tag: 'FAMS',
-              schemaOrgTag: null,
-              total: 32.39318102090577,
-              hasRDI: false,
-              daily: 0,
-              unit: 'g',
-            },
-            {
-              label: 'Polyunsaturated',
-              tag: 'FAPU',
-              schemaOrgTag: null,
-              total: 25.995804170774456,
-              hasRDI: false,
-              daily: 0,
-              unit: 'g',
-            },
-          ],
-        },
-        {
-          label: 'Carbs',
-          tag: 'CHOCDF',
-          schemaOrgTag: 'carbohydrateContent',
-          total: 47.87092499999999,
-          hasRDI: true,
-          daily: 15.956974999999996,
-          unit: 'g',
-          sub: [
-            {
-              label: 'Carbs (net)',
-              tag: 'CHOCDF.net',
-              schemaOrgTag: null,
-              total: 43.06392499999999,
-              hasRDI: false,
-              daily: 0,
-              unit: 'g',
-            },
-            {
-              label: 'Fiber',
-              tag: 'FIBTG',
-              schemaOrgTag: 'fiberContent',
-              total: 4.8069999999999995,
-              hasRDI: true,
-              daily: 19.227999999999998,
-              unit: 'g',
-            },
-            {
-              label: 'Sugars',
-              tag: 'SUGAR',
-              schemaOrgTag: 'sugarContent',
-              total: 8.915875000000002,
-              hasRDI: false,
-              daily: 0,
-              unit: 'g',
-            },
-            {
-              label: 'Sugars, added',
-              tag: 'SUGAR.added',
-              schemaOrgTag: null,
-              total: 0,
-              hasRDI: false,
-              daily: 0,
-              unit: 'g',
-            },
-          ],
-        },
-        {
-          label: 'Protein',
-          tag: 'PROCNT',
-          schemaOrgTag: 'proteinContent',
-          total: 74.69308489650001,
-          hasRDI: true,
-          daily: 149.38616979300002,
-          unit: 'g',
-        },
-        {
-          label: 'Cholesterol',
-          tag: 'CHOLE',
-          schemaOrgTag: 'cholesterolContent',
-          total: 235.2387702498096,
-          hasRDI: true,
-          daily: 78.4129234166032,
-          unit: 'mg',
-        },
-        {
-          label: 'Sodium',
-          tag: 'NA',
-          schemaOrgTag: 'sodiumContent',
-          total: 1515.910999147183,
-          hasRDI: true,
-          daily: 63.16295829779929,
-          unit: 'mg',
-        },
-        {
-          label: 'Calcium',
-          tag: 'CA',
-          schemaOrgTag: null,
-          total: 167.21850405,
-          hasRDI: true,
-          daily: 16.721850405,
-          unit: 'mg',
-        },
-        {
-          label: 'Magnesium',
-          tag: 'MG',
-          schemaOrgTag: null,
-          total: 120.2485844,
-          hasRDI: true,
-          daily: 28.630615333333335,
-          unit: 'mg',
-        },
-        {
-          label: 'Potassium',
-          tag: 'K',
-          schemaOrgTag: null,
-          total: 1273.5920087499999,
-          hasRDI: true,
-          daily: 27.09770231382978,
-          unit: 'mg',
-        },
-        {
-          label: 'Iron',
-          tag: 'FE',
-          schemaOrgTag: null,
-          total: 6.730529220000001,
-          hasRDI: true,
-          daily: 37.391829,
-          unit: 'mg',
-        },
-        {
-          label: 'Zinc',
-          tag: 'ZN',
-          schemaOrgTag: null,
-          total: 6.161625104500001,
-          hasRDI: true,
-          daily: 56.01477367727273,
-          unit: 'mg',
-        },
-        {
-          label: 'Phosphorus',
-          tag: 'P',
-          schemaOrgTag: null,
-          total: 747.2509041000001,
-          hasRDI: true,
-          daily: 106.75012915714287,
-          unit: 'mg',
-        },
-        {
-          label: 'Vitamin A',
-          tag: 'VITA_RAE',
-          schemaOrgTag: null,
-          total: 89.2409237,
-          hasRDI: true,
-          daily: 9.915658188888889,
-          unit: 'µg',
-        },
-        {
-          label: 'Vitamin C',
-          tag: 'VITC',
-          schemaOrgTag: null,
-          total: 13.552249999999999,
-          hasRDI: true,
-          daily: 15.058055555555555,
-          unit: 'mg',
-        },
-        {
-          label: 'Thiamin (B1)',
-          tag: 'THIA',
-          schemaOrgTag: null,
-          total: 0.8582842873,
-          hasRDI: true,
-          daily: 71.52369060833334,
-          unit: 'mg',
-        },
-        {
-          label: 'Riboflavin (B2)',
-          tag: 'RIBF',
-          schemaOrgTag: null,
-          total: 0.60481311275,
-          hasRDI: true,
-          daily: 46.524085596153846,
-          unit: 'mg',
-        },
-        {
-          label: 'Niacin (B3)',
-          tag: 'NIA',
-          schemaOrgTag: null,
-          total: 21.967686620000006,
-          hasRDI: true,
-          daily: 137.29804137500003,
-          unit: 'mg',
-        },
-        {
-          label: 'Vitamin B6',
-          tag: 'VITB6A',
-          schemaOrgTag: null,
-          total: 1.5839516880000002,
-          hasRDI: true,
-          daily: 121.84243753846155,
-          unit: 'mg',
-        },
-        {
-          label: 'Folate equivalent (total)',
-          tag: 'FOLDFE',
-          schemaOrgTag: null,
-          total: 130.44323295,
-          hasRDI: true,
-          daily: 32.6108082375,
-          unit: 'µg',
-        },
-        {
-          label: 'Folate (food)',
-          tag: 'FOLFD',
-          schemaOrgTag: null,
-          total: 100.86323295,
-          hasRDI: false,
-          daily: 0,
-          unit: 'µg',
-        },
-        {
-          label: 'Folic acid',
-          tag: 'FOLAC',
-          schemaOrgTag: null,
-          total: 17.4,
-          hasRDI: false,
-          daily: 0,
-          unit: 'µg',
-        },
-        {
-          label: 'Vitamin B12',
-          tag: 'VITB12',
-          schemaOrgTag: null,
-          total: 1.532543977,
-          hasRDI: true,
-          daily: 63.85599904166667,
-          unit: 'µg',
-        },
-        {
-          label: 'Vitamin D',
-          tag: 'VITD',
-          schemaOrgTag: null,
-          total: 0.46399999999999997,
-          hasRDI: true,
-          daily: 3.0933333333333333,
-          unit: 'µg',
-        },
-        {
-          label: 'Vitamin E',
-          tag: 'TOCPHA',
-          schemaOrgTag: null,
-          total: 1.19825,
-          hasRDI: true,
-          daily: 7.988333333333333,
-          unit: 'mg',
-        },
-        {
-          label: 'Vitamin K',
-          tag: 'VITK1',
-          schemaOrgTag: null,
-          total: 32.01075,
-          hasRDI: true,
-          daily: 26.675625000000004,
-          unit: 'µg',
-        },
-        {
-          label: 'Sugar alcohols',
-          tag: 'Sugar.alcohol',
-          schemaOrgTag: null,
-          total: 0,
-          hasRDI: false,
-          daily: 0,
-          unit: 'g',
-        },
-        {
-          label: 'Water',
-          tag: 'WATER',
-          schemaOrgTag: null,
-          total: 347.1338275925,
-          hasRDI: false,
-          daily: 0,
-          unit: 'g',
-        },
-      ],
-    },
-  },
   {
     recipe: {
       uri: 'http://www.edamam.com/ontologies/edamam.owl#recipe_e6f2680d029736fca03758483f43b336',
@@ -11681,6 +10237,1448 @@ const preloadedData = [
           tag: 'WATER',
           schemaOrgTag: null,
           total: 1.472659139078384,
+          hasRDI: false,
+          daily: 0,
+          unit: 'g',
+        },
+      ],
+    },
+  }, {
+    recipe: {
+      uri: 'http://www.edamam.com/ontologies/edamam.owl#recipe_e4456795dfe1409f69629df1670a6494',
+      label: 'Fruits Sandwich',
+      image: 'https://edamam-product-images.s3.amazonaws.com/web-img/5d7/5d77596f27841baeb6704eabb4a2fb5c.jpg?X-Amz-Security-Token=IQoJb3JpZ2luX2VjEE0aCXVzLWVhc3QtMSJIMEYCIQCjz1haqH%2BHU8PeUD5XPfPvhrg4B1tIA%2FchIC%2FdqMx%2FuAIhAPgRePErrrJffplWbN5DmRf%2BAVo4D9Hqt1Dj88GSmKD8KtUECKX%2F%2F%2F%2F%2F%2F%2F%2F%2F%2FwEQABoMMTg3MDE3MTUwOTg2Igyr1kArPKYLDQL7Tm4qqQS1JuqF6P2nantseLvzE7MeuUpQfcpjy%2BomgC7n9frUeX32ID6mQFtv6Q0vqUMOYVmLsS4xePgKLGKTLFvOoUUtpmmFiZy187eodT2KgAwjY%2BDU7WWfuYsnCl1PeWwkNTQEkD5O%2Bk3%2FjooadTuzHHLycTfvxJjpoxU5uyA5eYZtrqcPiA1XRG7ObdsEP%2FEuZYPG52mo%2F3oLWlWSpK41%2FrRQQkO4gKMHqriUvmo6YC02bi%2BSZhiAzV1R4C%2FaaW1NuAE%2FULEkSPsFbmyfBaureg6DinmQZlAj7XyP6vfMsUeZJJcs%2BW07eiylvgVi2SxqBWapl74KCeWUMsDZgobtyj6fVXltsjMAQgaZEY3dhmUipBfsXjAruutI0IcVguEIOremf4nSMM6bHGo1i%2BWcTMswT6HTEpDI0bQi%2Bi%2F0MBag8aa%2BE1CJn%2F3lbecKsngM3pvmqOvwAGLtfvzc6zwcBaW6fk%2BIzGXy%2FapYMI3LJjcSZremPwqnUhnjbLtL8DXvN09d1PT8DwXrtFzLkr5%2FervOO8A452AGZniBMqw2Sg0YoZ2Clu8PtxbUAuSy8asYI7Yeu3dsKor%2B1aqkr7b%2BAqALTpNUAuYsE7eZQKR2L62uoUFLu8BNfF5vNsiJV3iLMy%2BCdSAr2zMVnhvUPNK2Wkb0dfxZ7t89bznZVuH5ZBQ2Oh1TB9Ns8BjBD%2BMUaplrhya8IQXNk0npB4D3vHznNvjYVY4%2F%2FSiGZXYkMO7Q%2Bp0GOqgBhUfjEhUQQm93O5R9FZCA7T8ezNbjnvoWDlhSAl6MlnpXCgWuy3hIjNaGk60E4TPv0r7tLkyMiy28ZEtyMPt7iUka%2B2zJIr5pMDvbpFXEYYSpkOO1rl%2FoWE5uTgdEEN%2BWYm%2BtWMizTJIuFJ8e%2BzE%2BBkDxPQG7a3S%2B1pwp3OkKZ5k%2FnuvdyvQ2zlATQFxyFEhImfF3uo2y9RgpZfT8KooBqeL0tg%2FfUt3%2F&X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Date=20230111T131700Z&X-Amz-SignedHeaders=host&X-Amz-Expires=3600&X-Amz-Credential=ASIASXCYXIIFFLFRDZEE%2F20230111%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Signature=0b35c5a51686546c8247c938a7b2098db7aaea0f7f516512ed9d354ec16349e9',
+      source: 'Food52',
+      url: 'https://food52.com/recipes/61017-fruits-sandwich',
+      shareAs: 'http://www.edamam.com/recipe/fruits-sandwich-e4456795dfe1409f69629df1670a6494/sandwich',
+      yield: 2,
+      dietLabels: [
+        'High-Fiber',
+      ],
+      healthLabels: [],
+      cautions: [
+        'Sulfites',
+      ],
+      ingredientLines: [
+        '100 milliliters heavy cream',
+        '1 tablespoon sugar',
+        '4 pieces white sandwich bread',
+        '4 pieces strawberry',
+        '1 piece banana',
+        '2 pieces sliced kiwi',
+        '8 pieces tangerine',
+      ],
+      ingredients: [
+        {
+          text: '100 milliliters heavy cream',
+          quantity: 100,
+          measure: 'milliliter',
+          food: 'heavy cream',
+          weight: 100.76578765149213,
+          foodCategory: 'Dairy',
+          foodId: 'food_bgtkr21b5v16mca246x9ebnaswyo',
+          image: 'https://www.edamam.com/food-img/484/4848d71f6a14dd5076083f5e17925420.jpg',
+        },
+        {
+          text: '1 tablespoon sugar',
+          quantity: 1,
+          measure: 'tablespoon',
+          food: 'sugar',
+          weight: 12.4999999997887,
+          foodCategory: 'sugars',
+          foodId: 'food_axi2ijobrk819yb0adceobnhm1c2',
+          image: 'https://www.edamam.com/food-img/ecb/ecb3f5aaed96d0188c21b8369be07765.jpg',
+        },
+        {
+          text: '4 pieces white sandwich bread',
+          quantity: 4,
+          measure: 'piece',
+          food: 'sandwich bread',
+          weight: 116,
+          foodCategory: 'bread, rolls and tortillas',
+          foodId: 'food_a3049hmbqj5wstaeeb3udaz6uaqv',
+          image: 'https://www.edamam.com/food-img/886/886960f6ce6ccec5b9163bacf2996853.jpg',
+        },
+        {
+          text: '4 pieces strawberry',
+          quantity: 4,
+          measure: '<unit>',
+          food: 'strawberry',
+          weight: 53.733333333333334,
+          foodCategory: 'fruit',
+          foodId: 'food_b4s2ibkbrrucmbabbaxhfau8ay42',
+          image: 'https://www.edamam.com/food-img/00c/00c8851e401bf7975be7f73499b4b573.jpg',
+        },
+        {
+          text: '1 piece banana',
+          quantity: 1,
+          measure: '<unit>',
+          food: 'banana',
+          weight: 115.7,
+          foodCategory: 'fruit',
+          foodId: 'food_bjsfxtcaidvmhaa3afrbna43q3hu',
+          image: 'https://www.edamam.com/food-img/9f6/9f6181163a25c96022ee3fc66d9ebb11.jpg',
+        },
+        {
+          text: '2 pieces sliced kiwi',
+          quantity: 2,
+          measure: '<unit>',
+          food: 'kiwi',
+          weight: 138,
+          foodCategory: 'fruit',
+          foodId: 'food_bqj114aab4xnl6b6j3dt0bi4sznd',
+          image: 'https://www.edamam.com/food-img/dba/dbabbf58b6a43c9b65014f4c789d4409.jpg',
+        },
+        {
+          text: '8 pieces tangerine',
+          quantity: 8,
+          measure: 'piece',
+          food: 'tangerine',
+          weight: 226.4,
+          foodCategory: 'fruit',
+          foodId: 'food_a24i0asarijf6nbnhkbbrbgy91bw',
+          image: 'https://www.edamam.com/food-img/f82/f82193abe21e23c1562bb5340f74e336.jpg',
+        },
+      ],
+      calories: 1030.0766340634966,
+      totalWeight: 763.0991209846142,
+      totalTime: 15,
+      cuisineType: [
+        'british',
+      ],
+      mealType: [
+        'lunch/dinner',
+      ],
+      dishType: [
+        'sandwiches',
+      ],
+      totalNutrients: {
+        ENERC_KCAL: {
+          label: 'Energy',
+          quantity: 1030.0766340634966,
+          unit: 'kcal',
+        },
+        FAT: {
+          label: 'Fat',
+          quantity: 43.004191431052085,
+          unit: 'g',
+        },
+        FASAT: {
+          label: 'Saturated',
+          quantity: 24.375656211891666,
+          unit: 'g',
+        },
+        FATRN: {
+          label: 'Trans',
+          quantity: 0.030159999999999996,
+          unit: 'g',
+        },
+        FAMS: {
+          label: 'Monounsaturated',
+          quantity: 11.732781401771783,
+          unit: 'g',
+        },
+        FAPU: {
+          label: 'Polyunsaturated',
+          quantity: 3.8343295889981683,
+          unit: 'g',
+        },
+        CHOCDF: {
+          label: 'Carbs',
+          quantity: 152.76282547526537,
+          unit: 'g',
+        },
+        'CHOCDF.net': {
+          label: 'Carbohydrates (net)',
+          quantity: 135.82475880859872,
+          unit: 'g',
+        },
+        FIBTG: {
+          label: 'Fiber',
+          quantity: 16.938066666666668,
+          unit: 'g',
+        },
+        SUGAR: {
+          label: 'Sugars',
+          quantity: 75.27895547526575,
+          unit: 'g',
+        },
+        'SUGAR.added': {
+          label: 'Sugars, added',
+          quantity: 12.474999999789123,
+          unit: 'g',
+        },
+        PROCNT: {
+          label: 'Protein',
+          quantity: 19.52908198018892,
+          unit: 'g',
+        },
+        CHOLE: {
+          label: 'Cholesterol',
+          quantity: 138.04912908254423,
+          unit: 'mg',
+        },
+        NA: {
+          label: 'Sodium',
+          quantity: 638.0583326408982,
+          unit: 'mg',
+        },
+        CA: {
+          label: 'Calcium',
+          quantity: 364.9730953068011,
+          unit: 'mg',
+        },
+        MG: {
+          label: 'Magnesium',
+          quantity: 148.10593846893778,
+          unit: 'mg',
+        },
+        K: {
+          label: 'Potassium',
+          quantity: 1583.9463407386152,
+          unit: 'mg',
+        },
+        FE: {
+          label: 'Iron',
+          quantity: 5.373406402962009,
+          unit: 'mg',
+        },
+        ZN: {
+          label: 'Zinc',
+          quantity: 2.2138679782650774,
+          unit: 'mg',
+        },
+        P: {
+          label: 'Phosphorus',
+          quantity: 365.86478834392517,
+          unit: 'mg',
+        },
+        VITA_RAE: {
+          label: 'Vitamin A',
+          quantity: 500.65172058096596,
+          unit: 'µg',
+        },
+        VITC: {
+          label: 'Vitamin C',
+          quantity: 230.87249472590898,
+          unit: 'mg',
+        },
+        THIA: {
+          label: 'Thiamin (B1)',
+          quantity: 0.7209034732833282,
+          unit: 'mg',
+        },
+        RIBF: {
+          label: 'Riboflavin (B2)',
+          quantity: 0.6189836997499345,
+          unit: 'mg',
+        },
+        NIA: {
+          label: 'Niacin (B3)',
+          quantity: 8.857158323850749,
+          unit: 'mg',
+        },
+        VITB6A: {
+          label: 'Vitamin B6',
+          quantity: 0.8683647714560547,
+          unit: 'mg',
+        },
+        FOLDFE: {
+          label: 'Folate equivalent (total)',
+          quantity: 225.63063150605967,
+          unit: 'µg',
+        },
+        FOLFD: {
+          label: 'Folate (food)',
+          quantity: 186.19063150605967,
+          unit: 'µg',
+        },
+        FOLAC: {
+          label: 'Folic acid',
+          quantity: 23.2,
+          unit: 'µg',
+        },
+        VITB12: {
+          label: 'Vitamin B12',
+          quantity: 0.18137841777268585,
+          unit: 'µg',
+        },
+        VITD: {
+          label: 'Vitamin D',
+          quantity: 0.7053605135604449,
+          unit: 'µg',
+        },
+        TOCPHA: {
+          label: 'Vitamin E',
+          quantity: 4.027644015772483,
+          unit: 'mg',
+        },
+        VITK1: {
+          label: 'Vitamin K',
+          quantity: 66.28313853818108,
+          unit: 'µg',
+        },
+        'Sugar.alcohol': {
+          label: 'Sugar alcohol',
+          quantity: 0,
+          unit: 'g',
+        },
+        WATER: {
+          label: 'Water',
+          quantity: 541.9892527203427,
+          unit: 'g',
+        },
+      },
+      totalDaily: {
+        ENERC_KCAL: {
+          label: 'Energy',
+          quantity: 51.50383170317483,
+          unit: '%',
+        },
+        FAT: {
+          label: 'Fat',
+          quantity: 66.1602945093109,
+          unit: '%',
+        },
+        FASAT: {
+          label: 'Saturated',
+          quantity: 121.87828105945832,
+          unit: '%',
+        },
+        CHOCDF: {
+          label: 'Carbs',
+          quantity: 50.92094182508846,
+          unit: '%',
+        },
+        FIBTG: {
+          label: 'Fiber',
+          quantity: 67.75226666666667,
+          unit: '%',
+        },
+        PROCNT: {
+          label: 'Protein',
+          quantity: 39.05816396037784,
+          unit: '%',
+        },
+        CHOLE: {
+          label: 'Cholesterol',
+          quantity: 46.01637636084808,
+          unit: '%',
+        },
+        NA: {
+          label: 'Sodium',
+          quantity: 26.585763860037424,
+          unit: '%',
+        },
+        CA: {
+          label: 'Calcium',
+          quantity: 36.49730953068011,
+          unit: '%',
+        },
+        MG: {
+          label: 'Magnesium',
+          quantity: 35.26331868308043,
+          unit: '%',
+        },
+        K: {
+          label: 'Potassium',
+          quantity: 33.70098597316203,
+          unit: '%',
+        },
+        FE: {
+          label: 'Iron',
+          quantity: 29.85225779423338,
+          unit: '%',
+        },
+        ZN: {
+          label: 'Zinc',
+          quantity: 20.12607252968252,
+          unit: '%',
+        },
+        P: {
+          label: 'Phosphorus',
+          quantity: 52.26639833484646,
+          unit: '%',
+        },
+        VITA_RAE: {
+          label: 'Vitamin A',
+          quantity: 55.62796895344066,
+          unit: '%',
+        },
+        VITC: {
+          label: 'Vitamin C',
+          quantity: 256.52499413989887,
+          unit: '%',
+        },
+        THIA: {
+          label: 'Thiamin (B1)',
+          quantity: 60.075289440277345,
+          unit: '%',
+        },
+        RIBF: {
+          label: 'Riboflavin (B2)',
+          quantity: 47.61413074999496,
+          unit: '%',
+        },
+        NIA: {
+          label: 'Niacin (B3)',
+          quantity: 55.35723952406718,
+          unit: '%',
+        },
+        VITB6A: {
+          label: 'Vitamin B6',
+          quantity: 66.7972901120042,
+          unit: '%',
+        },
+        FOLDFE: {
+          label: 'Folate equivalent (total)',
+          quantity: 56.40765787651491,
+          unit: '%',
+        },
+        VITB12: {
+          label: 'Vitamin B12',
+          quantity: 7.55743407386191,
+          unit: '%',
+        },
+        VITD: {
+          label: 'Vitamin D',
+          quantity: 4.7024034237363,
+          unit: '%',
+        },
+        TOCPHA: {
+          label: 'Vitamin E',
+          quantity: 26.85096010514989,
+          unit: '%',
+        },
+        VITK1: {
+          label: 'Vitamin K',
+          quantity: 55.23594878181756,
+          unit: '%',
+        },
+      },
+      digest: [
+        {
+          label: 'Fat',
+          tag: 'FAT',
+          schemaOrgTag: 'fatContent',
+          total: 43.004191431052085,
+          hasRDI: true,
+          daily: 66.1602945093109,
+          unit: 'g',
+          sub: [
+            {
+              label: 'Saturated',
+              tag: 'FASAT',
+              schemaOrgTag: 'saturatedFatContent',
+              total: 24.375656211891666,
+              hasRDI: true,
+              daily: 121.87828105945832,
+              unit: 'g',
+            },
+            {
+              label: 'Trans',
+              tag: 'FATRN',
+              schemaOrgTag: 'transFatContent',
+              total: 0.030159999999999996,
+              hasRDI: false,
+              daily: 0,
+              unit: 'g',
+            },
+            {
+              label: 'Monounsaturated',
+              tag: 'FAMS',
+              schemaOrgTag: null,
+              total: 11.732781401771783,
+              hasRDI: false,
+              daily: 0,
+              unit: 'g',
+            },
+            {
+              label: 'Polyunsaturated',
+              tag: 'FAPU',
+              schemaOrgTag: null,
+              total: 3.8343295889981683,
+              hasRDI: false,
+              daily: 0,
+              unit: 'g',
+            },
+          ],
+        },
+        {
+          label: 'Carbs',
+          tag: 'CHOCDF',
+          schemaOrgTag: 'carbohydrateContent',
+          total: 152.76282547526537,
+          hasRDI: true,
+          daily: 50.92094182508846,
+          unit: 'g',
+          sub: [
+            {
+              label: 'Carbs (net)',
+              tag: 'CHOCDF.net',
+              schemaOrgTag: null,
+              total: 135.82475880859872,
+              hasRDI: false,
+              daily: 0,
+              unit: 'g',
+            },
+            {
+              label: 'Fiber',
+              tag: 'FIBTG',
+              schemaOrgTag: 'fiberContent',
+              total: 16.938066666666668,
+              hasRDI: true,
+              daily: 67.75226666666667,
+              unit: 'g',
+            },
+            {
+              label: 'Sugars',
+              tag: 'SUGAR',
+              schemaOrgTag: 'sugarContent',
+              total: 75.27895547526575,
+              hasRDI: false,
+              daily: 0,
+              unit: 'g',
+            },
+            {
+              label: 'Sugars, added',
+              tag: 'SUGAR.added',
+              schemaOrgTag: null,
+              total: 12.474999999789123,
+              hasRDI: false,
+              daily: 0,
+              unit: 'g',
+            },
+          ],
+        },
+        {
+          label: 'Protein',
+          tag: 'PROCNT',
+          schemaOrgTag: 'proteinContent',
+          total: 19.52908198018892,
+          hasRDI: true,
+          daily: 39.05816396037784,
+          unit: 'g',
+        },
+        {
+          label: 'Cholesterol',
+          tag: 'CHOLE',
+          schemaOrgTag: 'cholesterolContent',
+          total: 138.04912908254423,
+          hasRDI: true,
+          daily: 46.01637636084808,
+          unit: 'mg',
+        },
+        {
+          label: 'Sodium',
+          tag: 'NA',
+          schemaOrgTag: 'sodiumContent',
+          total: 638.0583326408982,
+          hasRDI: true,
+          daily: 26.585763860037424,
+          unit: 'mg',
+        },
+        {
+          label: 'Calcium',
+          tag: 'CA',
+          schemaOrgTag: null,
+          total: 364.9730953068011,
+          hasRDI: true,
+          daily: 36.49730953068011,
+          unit: 'mg',
+        },
+        {
+          label: 'Magnesium',
+          tag: 'MG',
+          schemaOrgTag: null,
+          total: 148.10593846893778,
+          hasRDI: true,
+          daily: 35.26331868308043,
+          unit: 'mg',
+        },
+        {
+          label: 'Potassium',
+          tag: 'K',
+          schemaOrgTag: null,
+          total: 1583.9463407386152,
+          hasRDI: true,
+          daily: 33.70098597316203,
+          unit: 'mg',
+        },
+        {
+          label: 'Iron',
+          tag: 'FE',
+          schemaOrgTag: null,
+          total: 5.373406402962009,
+          hasRDI: true,
+          daily: 29.85225779423338,
+          unit: 'mg',
+        },
+        {
+          label: 'Zinc',
+          tag: 'ZN',
+          schemaOrgTag: null,
+          total: 2.2138679782650774,
+          hasRDI: true,
+          daily: 20.12607252968252,
+          unit: 'mg',
+        },
+        {
+          label: 'Phosphorus',
+          tag: 'P',
+          schemaOrgTag: null,
+          total: 365.86478834392517,
+          hasRDI: true,
+          daily: 52.26639833484646,
+          unit: 'mg',
+        },
+        {
+          label: 'Vitamin A',
+          tag: 'VITA_RAE',
+          schemaOrgTag: null,
+          total: 500.65172058096596,
+          hasRDI: true,
+          daily: 55.62796895344066,
+          unit: 'µg',
+        },
+        {
+          label: 'Vitamin C',
+          tag: 'VITC',
+          schemaOrgTag: null,
+          total: 230.87249472590898,
+          hasRDI: true,
+          daily: 256.52499413989887,
+          unit: 'mg',
+        },
+        {
+          label: 'Thiamin (B1)',
+          tag: 'THIA',
+          schemaOrgTag: null,
+          total: 0.7209034732833282,
+          hasRDI: true,
+          daily: 60.075289440277345,
+          unit: 'mg',
+        },
+        {
+          label: 'Riboflavin (B2)',
+          tag: 'RIBF',
+          schemaOrgTag: null,
+          total: 0.6189836997499345,
+          hasRDI: true,
+          daily: 47.61413074999496,
+          unit: 'mg',
+        },
+        {
+          label: 'Niacin (B3)',
+          tag: 'NIA',
+          schemaOrgTag: null,
+          total: 8.857158323850749,
+          hasRDI: true,
+          daily: 55.35723952406718,
+          unit: 'mg',
+        },
+        {
+          label: 'Vitamin B6',
+          tag: 'VITB6A',
+          schemaOrgTag: null,
+          total: 0.8683647714560547,
+          hasRDI: true,
+          daily: 66.7972901120042,
+          unit: 'mg',
+        },
+        {
+          label: 'Folate equivalent (total)',
+          tag: 'FOLDFE',
+          schemaOrgTag: null,
+          total: 225.63063150605967,
+          hasRDI: true,
+          daily: 56.40765787651491,
+          unit: 'µg',
+        },
+        {
+          label: 'Folate (food)',
+          tag: 'FOLFD',
+          schemaOrgTag: null,
+          total: 186.19063150605967,
+          hasRDI: false,
+          daily: 0,
+          unit: 'µg',
+        },
+        {
+          label: 'Folic acid',
+          tag: 'FOLAC',
+          schemaOrgTag: null,
+          total: 23.2,
+          hasRDI: false,
+          daily: 0,
+          unit: 'µg',
+        },
+        {
+          label: 'Vitamin B12',
+          tag: 'VITB12',
+          schemaOrgTag: null,
+          total: 0.18137841777268585,
+          hasRDI: true,
+          daily: 7.55743407386191,
+          unit: 'µg',
+        },
+        {
+          label: 'Vitamin D',
+          tag: 'VITD',
+          schemaOrgTag: null,
+          total: 0.7053605135604449,
+          hasRDI: true,
+          daily: 4.7024034237363,
+          unit: 'µg',
+        },
+        {
+          label: 'Vitamin E',
+          tag: 'TOCPHA',
+          schemaOrgTag: null,
+          total: 4.027644015772483,
+          hasRDI: true,
+          daily: 26.85096010514989,
+          unit: 'mg',
+        },
+        {
+          label: 'Vitamin K',
+          tag: 'VITK1',
+          schemaOrgTag: null,
+          total: 66.28313853818108,
+          hasRDI: true,
+          daily: 55.23594878181756,
+          unit: 'µg',
+        },
+        {
+          label: 'Sugar alcohols',
+          tag: 'Sugar.alcohol',
+          schemaOrgTag: null,
+          total: 0,
+          hasRDI: false,
+          daily: 0,
+          unit: 'g',
+        },
+        {
+          label: 'Water',
+          tag: 'WATER',
+          schemaOrgTag: null,
+          total: 541.9892527203427,
+          hasRDI: false,
+          daily: 0,
+          unit: 'g',
+        },
+      ],
+    },
+  },
+  {
+    recipe: {
+      uri: 'http://www.edamam.com/ontologies/edamam.owl#recipe_2256482f9dc843a41d747ca9d76690e5',
+      label: 'Turkey Club Sandwich recipes',
+      image: 'https://edamam-product-images.s3.amazonaws.com/web-img/ab5/ab5ddc22f4c01397fca3c24e2dab0617?X-Amz-Security-Token=IQoJb3JpZ2luX2VjEE0aCXVzLWVhc3QtMSJIMEYCIQCjz1haqH%2BHU8PeUD5XPfPvhrg4B1tIA%2FchIC%2FdqMx%2FuAIhAPgRePErrrJffplWbN5DmRf%2BAVo4D9Hqt1Dj88GSmKD8KtUECKX%2F%2F%2F%2F%2F%2F%2F%2F%2F%2FwEQABoMMTg3MDE3MTUwOTg2Igyr1kArPKYLDQL7Tm4qqQS1JuqF6P2nantseLvzE7MeuUpQfcpjy%2BomgC7n9frUeX32ID6mQFtv6Q0vqUMOYVmLsS4xePgKLGKTLFvOoUUtpmmFiZy187eodT2KgAwjY%2BDU7WWfuYsnCl1PeWwkNTQEkD5O%2Bk3%2FjooadTuzHHLycTfvxJjpoxU5uyA5eYZtrqcPiA1XRG7ObdsEP%2FEuZYPG52mo%2F3oLWlWSpK41%2FrRQQkO4gKMHqriUvmo6YC02bi%2BSZhiAzV1R4C%2FaaW1NuAE%2FULEkSPsFbmyfBaureg6DinmQZlAj7XyP6vfMsUeZJJcs%2BW07eiylvgVi2SxqBWapl74KCeWUMsDZgobtyj6fVXltsjMAQgaZEY3dhmUipBfsXjAruutI0IcVguEIOremf4nSMM6bHGo1i%2BWcTMswT6HTEpDI0bQi%2Bi%2F0MBag8aa%2BE1CJn%2F3lbecKsngM3pvmqOvwAGLtfvzc6zwcBaW6fk%2BIzGXy%2FapYMI3LJjcSZremPwqnUhnjbLtL8DXvN09d1PT8DwXrtFzLkr5%2FervOO8A452AGZniBMqw2Sg0YoZ2Clu8PtxbUAuSy8asYI7Yeu3dsKor%2B1aqkr7b%2BAqALTpNUAuYsE7eZQKR2L62uoUFLu8BNfF5vNsiJV3iLMy%2BCdSAr2zMVnhvUPNK2Wkb0dfxZ7t89bznZVuH5ZBQ2Oh1TB9Ns8BjBD%2BMUaplrhya8IQXNk0npB4D3vHznNvjYVY4%2F%2FSiGZXYkMO7Q%2Bp0GOqgBhUfjEhUQQm93O5R9FZCA7T8ezNbjnvoWDlhSAl6MlnpXCgWuy3hIjNaGk60E4TPv0r7tLkyMiy28ZEtyMPt7iUka%2B2zJIr5pMDvbpFXEYYSpkOO1rl%2FoWE5uTgdEEN%2BWYm%2BtWMizTJIuFJ8e%2BzE%2BBkDxPQG7a3S%2B1pwp3OkKZ5k%2FnuvdyvQ2zlATQFxyFEhImfF3uo2y9RgpZfT8KooBqeL0tg%2FfUt3%2F&X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Date=20230111T131700Z&X-Amz-SignedHeaders=host&X-Amz-Expires=3600&X-Amz-Credential=ASIASXCYXIIFFLFRDZEE%2F20230111%2Fus-east-1%2Fs3%2Faws4_request&X-Amz-Signature=36e24c9d9e86383017fae3d366ea0bdcf79f46668d2f8a0cd7e958c41c085b69',
+      source: 'Food Republic',
+      url: 'http://www.foodrepublic.com/recipes/turkey-club-sandwich-recipe/',
+      shareAs: 'http://www.edamam.com/recipe/turkey-club-sandwich-recipes-2256482f9dc843a41d747ca9d76690e5/sandwich',
+      yield: 4,
+      dietLabels: [
+        'Low-Carb',
+      ],
+      healthLabels: [],
+      cautions: [
+        'Sulfites',
+      ],
+      ingredientLines: [
+        '3 slices white sandwich bread, lightly toasted',
+        '2 tablespoons mayonnaise',
+        '1/2 pound turkey breast, thinly sliced',
+        '4 slices thick-cut bacon, fried until crisp',
+        '1 small tomato, sliced into rounds',
+        '4 leaves lettuce, rinsed and dried',
+      ],
+      ingredients: [
+        {
+          text: '3 slices white sandwich bread, lightly toasted',
+          quantity: 3,
+          measure: 'slice',
+          food: 'sandwich bread',
+          weight: 87,
+          foodCategory: 'bread, rolls and tortillas',
+          foodId: 'food_a3049hmbqj5wstaeeb3udaz6uaqv',
+          image: 'https://www.edamam.com/food-img/886/886960f6ce6ccec5b9163bacf2996853.jpg',
+        },
+        {
+          text: '2 tablespoons mayonnaise',
+          quantity: 2,
+          measure: 'tablespoon',
+          food: 'mayonnaise',
+          weight: 28.8749999995118,
+          foodCategory: 'condiments and sauces',
+          foodId: 'food_bu8t61zaplle7dbrzk81dbygq0qj',
+          image: 'https://www.edamam.com/food-img/577/577308a0422357885c94cc9b5f1f1862.jpg',
+        },
+        {
+          text: '1/2 pound turkey breast, thinly sliced',
+          quantity: 0.5,
+          measure: 'pound',
+          food: 'turkey breast',
+          weight: 226.796185,
+          foodCategory: 'Poultry',
+          foodId: 'food_aijy334arptsr3b4zwpdrajqapl8',
+          image: 'https://www.edamam.com/food-img/3d8/3d80f3f4fcb0efcf54d094c857d74926.jpg',
+        },
+        {
+          text: '4 slices thick-cut bacon, fried until crisp',
+          quantity: 4,
+          measure: 'slice',
+          food: 'bacon',
+          weight: 116,
+          foodCategory: 'cured meats',
+          foodId: 'food_av0awb0bgs18qub01sd9kbx44b61',
+          image: 'https://www.edamam.com/food-img/d42/d426884a125fa39a70d5a5d7217864ec.jpg',
+        },
+        {
+          text: '1 small tomato, sliced into rounds',
+          quantity: 1,
+          measure: '<unit>',
+          food: 'tomato',
+          weight: 92.25,
+          foodCategory: 'vegetables',
+          foodId: 'food_a6k79rrahp8fe2b26zussa3wtkqh',
+          image: 'https://www.edamam.com/food-img/23e/23e727a14f1035bdc2733bb0477efbd2.jpg',
+        },
+        {
+          text: '4 leaves lettuce, rinsed and dried',
+          quantity: 4,
+          measure: 'leaf',
+          food: 'lettuce',
+          weight: 20,
+          foodCategory: 'vegetables',
+          foodId: 'food_bf5fxtkbc9alwoajuvsi7amonr5w',
+          image: 'https://www.edamam.com/food-img/719/71996625d0cb47e197093ecd52c97dc2.jpg',
+        },
+      ],
+      calories: 1291.1000104466216,
+      totalWeight: 570.9211849995118,
+      totalTime: 15,
+      cuisineType: [
+        'american',
+      ],
+      mealType: [
+        'lunch/dinner',
+      ],
+      dishType: [
+        'sandwiches',
+      ],
+      totalNutrients: {
+        ENERC_KCAL: {
+          label: 'Energy',
+          quantity: 1291.1000104466216,
+          unit: 'kcal',
+        },
+        FAT: {
+          label: 'Fat',
+          quantity: 87.24254218662408,
+          unit: 'g',
+        },
+        FASAT: {
+          label: 'Saturated',
+          quantity: 23.783412133443857,
+          unit: 'g',
+        },
+        FATRN: {
+          label: 'Trans',
+          quantity: 0.1769,
+          unit: 'g',
+        },
+        FAMS: {
+          label: 'Monounsaturated',
+          quantity: 32.39318102090577,
+          unit: 'g',
+        },
+        FAPU: {
+          label: 'Polyunsaturated',
+          quantity: 25.995804170774456,
+          unit: 'g',
+        },
+        CHOCDF: {
+          label: 'Carbs',
+          quantity: 47.87092499999999,
+          unit: 'g',
+        },
+        'CHOCDF.net': {
+          label: 'Carbohydrates (net)',
+          quantity: 43.06392499999999,
+          unit: 'g',
+        },
+        FIBTG: {
+          label: 'Fiber',
+          quantity: 4.8069999999999995,
+          unit: 'g',
+        },
+        SUGAR: {
+          label: 'Sugars',
+          quantity: 8.915875000000002,
+          unit: 'g',
+        },
+        'SUGAR.added': {
+          label: 'Sugars, added',
+          quantity: 0,
+          unit: 'g',
+        },
+        PROCNT: {
+          label: 'Protein',
+          quantity: 74.69308489650001,
+          unit: 'g',
+        },
+        CHOLE: {
+          label: 'Cholesterol',
+          quantity: 235.2387702498096,
+          unit: 'mg',
+        },
+        NA: {
+          label: 'Sodium',
+          quantity: 1515.910999147183,
+          unit: 'mg',
+        },
+        CA: {
+          label: 'Calcium',
+          quantity: 167.21850405,
+          unit: 'mg',
+        },
+        MG: {
+          label: 'Magnesium',
+          quantity: 120.2485844,
+          unit: 'mg',
+        },
+        K: {
+          label: 'Potassium',
+          quantity: 1273.5920087499999,
+          unit: 'mg',
+        },
+        FE: {
+          label: 'Iron',
+          quantity: 6.730529220000001,
+          unit: 'mg',
+        },
+        ZN: {
+          label: 'Zinc',
+          quantity: 6.161625104500001,
+          unit: 'mg',
+        },
+        P: {
+          label: 'Phosphorus',
+          quantity: 747.2509041000001,
+          unit: 'mg',
+        },
+        VITA_RAE: {
+          label: 'Vitamin A',
+          quantity: 89.2409237,
+          unit: 'µg',
+        },
+        VITC: {
+          label: 'Vitamin C',
+          quantity: 13.552249999999999,
+          unit: 'mg',
+        },
+        THIA: {
+          label: 'Thiamin (B1)',
+          quantity: 0.8582842873,
+          unit: 'mg',
+        },
+        RIBF: {
+          label: 'Riboflavin (B2)',
+          quantity: 0.60481311275,
+          unit: 'mg',
+        },
+        NIA: {
+          label: 'Niacin (B3)',
+          quantity: 21.967686620000006,
+          unit: 'mg',
+        },
+        VITB6A: {
+          label: 'Vitamin B6',
+          quantity: 1.5839516880000002,
+          unit: 'mg',
+        },
+        FOLDFE: {
+          label: 'Folate equivalent (total)',
+          quantity: 130.44323295,
+          unit: 'µg',
+        },
+        FOLFD: {
+          label: 'Folate (food)',
+          quantity: 100.86323295,
+          unit: 'µg',
+        },
+        FOLAC: {
+          label: 'Folic acid',
+          quantity: 17.4,
+          unit: 'µg',
+        },
+        VITB12: {
+          label: 'Vitamin B12',
+          quantity: 1.532543977,
+          unit: 'µg',
+        },
+        VITD: {
+          label: 'Vitamin D',
+          quantity: 0.46399999999999997,
+          unit: 'µg',
+        },
+        TOCPHA: {
+          label: 'Vitamin E',
+          quantity: 1.19825,
+          unit: 'mg',
+        },
+        VITK1: {
+          label: 'Vitamin K',
+          quantity: 32.01075,
+          unit: 'µg',
+        },
+        'Sugar.alcohol': {
+          label: 'Sugar alcohol',
+          quantity: 0,
+          unit: 'g',
+        },
+        WATER: {
+          label: 'Water',
+          quantity: 347.1338275925,
+          unit: 'g',
+        },
+      },
+      totalDaily: {
+        ENERC_KCAL: {
+          label: 'Energy',
+          quantity: 64.55500052233108,
+          unit: '%',
+        },
+        FAT: {
+          label: 'Fat',
+          quantity: 134.21929567172936,
+          unit: '%',
+        },
+        FASAT: {
+          label: 'Saturated',
+          quantity: 118.91706066721929,
+          unit: '%',
+        },
+        CHOCDF: {
+          label: 'Carbs',
+          quantity: 15.956974999999996,
+          unit: '%',
+        },
+        FIBTG: {
+          label: 'Fiber',
+          quantity: 19.227999999999998,
+          unit: '%',
+        },
+        PROCNT: {
+          label: 'Protein',
+          quantity: 149.38616979300002,
+          unit: '%',
+        },
+        CHOLE: {
+          label: 'Cholesterol',
+          quantity: 78.4129234166032,
+          unit: '%',
+        },
+        NA: {
+          label: 'Sodium',
+          quantity: 63.16295829779929,
+          unit: '%',
+        },
+        CA: {
+          label: 'Calcium',
+          quantity: 16.721850405,
+          unit: '%',
+        },
+        MG: {
+          label: 'Magnesium',
+          quantity: 28.630615333333335,
+          unit: '%',
+        },
+        K: {
+          label: 'Potassium',
+          quantity: 27.09770231382978,
+          unit: '%',
+        },
+        FE: {
+          label: 'Iron',
+          quantity: 37.391829,
+          unit: '%',
+        },
+        ZN: {
+          label: 'Zinc',
+          quantity: 56.01477367727273,
+          unit: '%',
+        },
+        P: {
+          label: 'Phosphorus',
+          quantity: 106.75012915714287,
+          unit: '%',
+        },
+        VITA_RAE: {
+          label: 'Vitamin A',
+          quantity: 9.915658188888889,
+          unit: '%',
+        },
+        VITC: {
+          label: 'Vitamin C',
+          quantity: 15.058055555555555,
+          unit: '%',
+        },
+        THIA: {
+          label: 'Thiamin (B1)',
+          quantity: 71.52369060833334,
+          unit: '%',
+        },
+        RIBF: {
+          label: 'Riboflavin (B2)',
+          quantity: 46.524085596153846,
+          unit: '%',
+        },
+        NIA: {
+          label: 'Niacin (B3)',
+          quantity: 137.29804137500003,
+          unit: '%',
+        },
+        VITB6A: {
+          label: 'Vitamin B6',
+          quantity: 121.84243753846155,
+          unit: '%',
+        },
+        FOLDFE: {
+          label: 'Folate equivalent (total)',
+          quantity: 32.6108082375,
+          unit: '%',
+        },
+        VITB12: {
+          label: 'Vitamin B12',
+          quantity: 63.85599904166667,
+          unit: '%',
+        },
+        VITD: {
+          label: 'Vitamin D',
+          quantity: 3.0933333333333333,
+          unit: '%',
+        },
+        TOCPHA: {
+          label: 'Vitamin E',
+          quantity: 7.988333333333333,
+          unit: '%',
+        },
+        VITK1: {
+          label: 'Vitamin K',
+          quantity: 26.675625000000004,
+          unit: '%',
+        },
+      },
+      digest: [
+        {
+          label: 'Fat',
+          tag: 'FAT',
+          schemaOrgTag: 'fatContent',
+          total: 87.24254218662408,
+          hasRDI: true,
+          daily: 134.21929567172936,
+          unit: 'g',
+          sub: [
+            {
+              label: 'Saturated',
+              tag: 'FASAT',
+              schemaOrgTag: 'saturatedFatContent',
+              total: 23.783412133443857,
+              hasRDI: true,
+              daily: 118.91706066721929,
+              unit: 'g',
+            },
+            {
+              label: 'Trans',
+              tag: 'FATRN',
+              schemaOrgTag: 'transFatContent',
+              total: 0.1769,
+              hasRDI: false,
+              daily: 0,
+              unit: 'g',
+            },
+            {
+              label: 'Monounsaturated',
+              tag: 'FAMS',
+              schemaOrgTag: null,
+              total: 32.39318102090577,
+              hasRDI: false,
+              daily: 0,
+              unit: 'g',
+            },
+            {
+              label: 'Polyunsaturated',
+              tag: 'FAPU',
+              schemaOrgTag: null,
+              total: 25.995804170774456,
+              hasRDI: false,
+              daily: 0,
+              unit: 'g',
+            },
+          ],
+        },
+        {
+          label: 'Carbs',
+          tag: 'CHOCDF',
+          schemaOrgTag: 'carbohydrateContent',
+          total: 47.87092499999999,
+          hasRDI: true,
+          daily: 15.956974999999996,
+          unit: 'g',
+          sub: [
+            {
+              label: 'Carbs (net)',
+              tag: 'CHOCDF.net',
+              schemaOrgTag: null,
+              total: 43.06392499999999,
+              hasRDI: false,
+              daily: 0,
+              unit: 'g',
+            },
+            {
+              label: 'Fiber',
+              tag: 'FIBTG',
+              schemaOrgTag: 'fiberContent',
+              total: 4.8069999999999995,
+              hasRDI: true,
+              daily: 19.227999999999998,
+              unit: 'g',
+            },
+            {
+              label: 'Sugars',
+              tag: 'SUGAR',
+              schemaOrgTag: 'sugarContent',
+              total: 8.915875000000002,
+              hasRDI: false,
+              daily: 0,
+              unit: 'g',
+            },
+            {
+              label: 'Sugars, added',
+              tag: 'SUGAR.added',
+              schemaOrgTag: null,
+              total: 0,
+              hasRDI: false,
+              daily: 0,
+              unit: 'g',
+            },
+          ],
+        },
+        {
+          label: 'Protein',
+          tag: 'PROCNT',
+          schemaOrgTag: 'proteinContent',
+          total: 74.69308489650001,
+          hasRDI: true,
+          daily: 149.38616979300002,
+          unit: 'g',
+        },
+        {
+          label: 'Cholesterol',
+          tag: 'CHOLE',
+          schemaOrgTag: 'cholesterolContent',
+          total: 235.2387702498096,
+          hasRDI: true,
+          daily: 78.4129234166032,
+          unit: 'mg',
+        },
+        {
+          label: 'Sodium',
+          tag: 'NA',
+          schemaOrgTag: 'sodiumContent',
+          total: 1515.910999147183,
+          hasRDI: true,
+          daily: 63.16295829779929,
+          unit: 'mg',
+        },
+        {
+          label: 'Calcium',
+          tag: 'CA',
+          schemaOrgTag: null,
+          total: 167.21850405,
+          hasRDI: true,
+          daily: 16.721850405,
+          unit: 'mg',
+        },
+        {
+          label: 'Magnesium',
+          tag: 'MG',
+          schemaOrgTag: null,
+          total: 120.2485844,
+          hasRDI: true,
+          daily: 28.630615333333335,
+          unit: 'mg',
+        },
+        {
+          label: 'Potassium',
+          tag: 'K',
+          schemaOrgTag: null,
+          total: 1273.5920087499999,
+          hasRDI: true,
+          daily: 27.09770231382978,
+          unit: 'mg',
+        },
+        {
+          label: 'Iron',
+          tag: 'FE',
+          schemaOrgTag: null,
+          total: 6.730529220000001,
+          hasRDI: true,
+          daily: 37.391829,
+          unit: 'mg',
+        },
+        {
+          label: 'Zinc',
+          tag: 'ZN',
+          schemaOrgTag: null,
+          total: 6.161625104500001,
+          hasRDI: true,
+          daily: 56.01477367727273,
+          unit: 'mg',
+        },
+        {
+          label: 'Phosphorus',
+          tag: 'P',
+          schemaOrgTag: null,
+          total: 747.2509041000001,
+          hasRDI: true,
+          daily: 106.75012915714287,
+          unit: 'mg',
+        },
+        {
+          label: 'Vitamin A',
+          tag: 'VITA_RAE',
+          schemaOrgTag: null,
+          total: 89.2409237,
+          hasRDI: true,
+          daily: 9.915658188888889,
+          unit: 'µg',
+        },
+        {
+          label: 'Vitamin C',
+          tag: 'VITC',
+          schemaOrgTag: null,
+          total: 13.552249999999999,
+          hasRDI: true,
+          daily: 15.058055555555555,
+          unit: 'mg',
+        },
+        {
+          label: 'Thiamin (B1)',
+          tag: 'THIA',
+          schemaOrgTag: null,
+          total: 0.8582842873,
+          hasRDI: true,
+          daily: 71.52369060833334,
+          unit: 'mg',
+        },
+        {
+          label: 'Riboflavin (B2)',
+          tag: 'RIBF',
+          schemaOrgTag: null,
+          total: 0.60481311275,
+          hasRDI: true,
+          daily: 46.524085596153846,
+          unit: 'mg',
+        },
+        {
+          label: 'Niacin (B3)',
+          tag: 'NIA',
+          schemaOrgTag: null,
+          total: 21.967686620000006,
+          hasRDI: true,
+          daily: 137.29804137500003,
+          unit: 'mg',
+        },
+        {
+          label: 'Vitamin B6',
+          tag: 'VITB6A',
+          schemaOrgTag: null,
+          total: 1.5839516880000002,
+          hasRDI: true,
+          daily: 121.84243753846155,
+          unit: 'mg',
+        },
+        {
+          label: 'Folate equivalent (total)',
+          tag: 'FOLDFE',
+          schemaOrgTag: null,
+          total: 130.44323295,
+          hasRDI: true,
+          daily: 32.6108082375,
+          unit: 'µg',
+        },
+        {
+          label: 'Folate (food)',
+          tag: 'FOLFD',
+          schemaOrgTag: null,
+          total: 100.86323295,
+          hasRDI: false,
+          daily: 0,
+          unit: 'µg',
+        },
+        {
+          label: 'Folic acid',
+          tag: 'FOLAC',
+          schemaOrgTag: null,
+          total: 17.4,
+          hasRDI: false,
+          daily: 0,
+          unit: 'µg',
+        },
+        {
+          label: 'Vitamin B12',
+          tag: 'VITB12',
+          schemaOrgTag: null,
+          total: 1.532543977,
+          hasRDI: true,
+          daily: 63.85599904166667,
+          unit: 'µg',
+        },
+        {
+          label: 'Vitamin D',
+          tag: 'VITD',
+          schemaOrgTag: null,
+          total: 0.46399999999999997,
+          hasRDI: true,
+          daily: 3.0933333333333333,
+          unit: 'µg',
+        },
+        {
+          label: 'Vitamin E',
+          tag: 'TOCPHA',
+          schemaOrgTag: null,
+          total: 1.19825,
+          hasRDI: true,
+          daily: 7.988333333333333,
+          unit: 'mg',
+        },
+        {
+          label: 'Vitamin K',
+          tag: 'VITK1',
+          schemaOrgTag: null,
+          total: 32.01075,
+          hasRDI: true,
+          daily: 26.675625000000004,
+          unit: 'µg',
+        },
+        {
+          label: 'Sugar alcohols',
+          tag: 'Sugar.alcohol',
+          schemaOrgTag: null,
+          total: 0,
+          hasRDI: false,
+          daily: 0,
+          unit: 'g',
+        },
+        {
+          label: 'Water',
+          tag: 'WATER',
+          schemaOrgTag: null,
+          total: 347.1338275925,
           hasRDI: false,
           daily: 0,
           unit: 'g',
